mlops/cicd-pipeline/git-repo-codes/requirements.txt
sagemaker==2.73.0
stepfunctions==2.3.0
PyYAML==5.4.1
